Refactor tests for DraggableImage move
