+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SVGComponents/SVGImage/index.js
@@ -1,1 +1,2 @@
 export { default } from './SVGImage'
+export { DraggableImage } from './SVGImage'
